requirements-render.txt
Django==4.2.16
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
mongoengine==0.29.1
pymongo==4.6.1
dnspython==2.4.2
Pillow==10.1.0
requests==2.31.0
django-cors-headers==4.3.1
django-redis==5.4.0
redis==5.0.1
celery==5.3.4
django-extensions==3.2.3
django-debug-toolbar==4.2.0